block permissions for security features
Currently, features like sandboxing, impersonation, etc. are implemented
such that:

- view data data permissions for the relevant databases and tables are
`unrestricted`
- rows are created in other tables (e.g. `sandboxes`) that configure
sandboxing on the database or table
- at query time, EE-only code modifies the query in some way to create
the "special" behavior. For example, the sandboxing middleware replaces
the table being queried with a sandboxed sub-query.

Note that the enterprise code is *reducing* the permissions that are
defined in the data permissions table. The user starts out with
`unrestricted` access, and that access is amended to reduce their access
when they execute a query.

I am changing this behavior to something I think will be much safer:
- view data permissions for the relevant databases and tables are
`blocked`.
- rows are created in other tables (e.g. `sandboxes`) that configure
sandboxing on the database or table (same as now).
- at query time, we accept that a table may be blocked iff it is used
for sandboxing

Similarly, a database may be blocked iff it is used for impersonation.

In some ways, this change may be scary: we previously had an invariant
that if you are querying a table, you must have `unrestricted` data
access to it.

But I think the modified invariant is actually safer: if you are
querying a table, you must have `unrestricted` data access to it, OR it
must be accessed through a sandbox.

This should not have any effect on what people are currently allowed to
do, it just provides an additional safety harness.

As part of this change, I removed a "feature" of data permissions, where
we maintained an invariant:

- if you have `view-data: blocked` you must have `create-queries: no`.

- if you have `create-queries: !no`, you must have `view-data:
unrestricted`

We did this by modifying nominally separate permissions when another
permission was set. For example, if you set `create-queries` to a value
other than `no`, we would auto-set `view-data` to `unrestricted` at the
same time. Conversely, if you set `view-data` to `blocked`, we would
auto-set `create-queries` and `download-results` to `no`.

I've removed this logic. I feel comfortable about this because:

- In the long term, we actually want permissions to be orthogonal. It
should be meaningful to have `create-queries` permission, but NOT have
access to the database - you can create queries but not actually run
them.

- In the short term, there is no practical difference here. We check for
these permissions in tandem. Everywhere we check for `create-queries`
permission != `no`, we also check for `view-data` permission =
`unrestricted`.

- The FE duplicates the logic from the BE, and won't let you set
`create-queries` to non-`no` values unless the DB/schema/table is
un-blocked, and sets `create-queries` to `no` when the DB/schema/table
is blocked.

Theoretically if

- someone is directly using our API to set permissions,

- they are *explicitly* granting `create-queries` permissions in the
perm-graph they're passing,

- but are relying on the fact that those explicitly passed permissions
are not actually granted (because the `view-data` permission overrides
`create-queries`)

Permissions will get set differently in the database than before.

But as mentioned earlier, this won't actually have a functional effect,
so... I think it's fine.